+++ hutool-core/src/test/java/cn/hutool/core/lang/DictTest.java
@@ -1,12 +1,14 @@
 package cn.hutool.core.lang;
 
+import cn.hutool.core.builder.GenericBuilder;
 import cn.hutool.core.date.DateTime;
 import org.junit.Assert;
 import org.junit.Test;
-import static cn.hutool.core.lang.OptTest.User;
 
 import java.util.HashMap;
 import java.util.Map;
+
+import static cn.hutool.core.lang.OptTest.User;
 
 public class DictTest {
 	@Test
@@ -62,7 +64,7 @@ public class DictTest {
 
 	@Test
 	public void setFieldsTest() {
-		User user = User.builder().username("hutool").nickname(null).build();
+		User user = GenericBuilder.of(User::new).with(User::setUsername, "hutool").build();
 		Dict dict = Dict.create();
 		dict.setFields(user::getNickname, user::getUsername);
 		Assert.assertEquals("hutool", dict.get("username"));